Use StoreObservable protocol with protocol extension
