--- Source/Shared/Extensions/Dictionary+Extensions.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension Dictionary where Key == UUID {
-  mutating func insert(_ value: Value) -> UUID {
-    let id = UUID()
-    self[id] = value
-    return id
-  }
-}
